Fix print incorrectly removing messages (#18)
`print` was removing messages that were never set as IDs were set to 0 before - sending this through to A2's `removeMessage` was removing the first message even if that wasn't the ID returned.

Set all IDs to a default value (-1) and check against that before removing the message. Also tidied up the code a little and moved the functions to namespace level instead of being static functions in a class.

Added a new global function `clearPrint`. This removes all current print messages.

Closes #19