Enable the save/load buttons in FO SP
If loaded into FleetOps, alter the Transport::IsInstantAction function to always return true. This enables the save and load buttons in the menu in FleetOps as it doesn't seem to work with single player mode.

Also adds a couple more noexcept(false) in luabind to allow script errors.

Update to 1.0.4.

Closes #24